+++ extensions/clockify/src/index.tsx
@@ -1,10 +1,10 @@
-import { ActionPanel, Form, Icon, List, showToast, useNavigation, Toast, Action, LocalStorage } from "@raycast/api";
+import { ActionPanel, Form, Icon, List, showToast, useNavigation, Toast, Action, LocalStorage, openExtensionPreferences } from "@raycast/api";
 import { useCallback, useEffect, useState } from "react";
 import isEmpty from "lodash.isempty";
 import uniqWith from "lodash.uniqwith";
 import useConfig from "./useConfig";
 import { fetcher, isInProgress, showElapsedTime } from "./utils";
-import { TimeEntry, Project, Task } from "./types";
+import { TimeEntry, Project, Task, Tag } from "./types";
 
 function OpenWebPage() {
   return <Action.OpenInBrowser title="Open Website" url="https://app.clockify.me" />;
@@ -36,7 +36,7 @@ function ItemInProgress({ entry, updateTimeEntries }: { entry: TimeEntry; update
       keywords={[...(entry.description?.split(" ") ?? []), ...(entry.project?.name.split(" ") ?? [])]}
       actions={
         <ActionPanel>
-          <Action title="Stop Timer" onAction={() => stopCurrentTimer().then(() => updateTimeEntries())} />
+          <Action icon={Icon.Stop} title="Stop Timer" onAction={() => stopCurrentTimer().then(() => updateTimeEntries())} />
           <OpenWebPage />
         </ActionPanel>
       }
@@ -48,7 +48,6 @@ export default function Main() {
   const { config, isValidToken, setIsValidToken } = useConfig();
   const [entries, setEntries] = useState<TimeEntry[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { push } = useNavigation();
 
   useEffect(() => {
     if (isEmpty(config) || !isValidToken) return;
@@ -97,6 +96,9 @@ export default function Main() {
           icon={Icon.ExclamationMark}
           title="Invalid API Key Detected"
           accessories={[{ text: `Go to Extensions → Clockify` }]}
+          actions={<ActionPanel>
+            <Action icon={Icon.Gear} title="Open Extension Preferences" onAction={openExtensionPreferences} />
+          </ActionPanel>}
         />
       ) : (
         <>
@@ -106,9 +108,10 @@ export default function Main() {
               title="Start New Timer"
               actions={
                 <ActionPanel>
-                  <Action
+                  <Action.Push
+                    icon={Icon.ArrowRight}
                     title="Start New Timer"
-                    onAction={() => push(<NewEntry updateTimeEntries={updateTimeEntries} />)}
+                    target={<NewEntry updateTimeEntries={updateTimeEntries} />}
                   />
                   <OpenWebPage />
                 </ActionPanel>
@@ -127,12 +130,14 @@ export default function Main() {
                   subtitle={`${[entry.description || "No Description", entry.task?.name].filter(Boolean).join(" • ")}`}
                   accessories={[
                     { text: entry.project?.name, icon: { source: Icon.Dot, tintColor: entry.project?.color } },
+                    {text: entry.tags.length}
                   ]}
                   icon={{ source: Icon.Circle, tintColor: entry.project?.color }}
                   keywords={[...(entry.description?.split(" ") ?? []), ...(entry.project?.name.split(" ") ?? [])]}
                   actions={
                     <ActionPanel>
                       <Action
+                        icon={Icon.Play}
                         title="Start Timer"
                         onAction={() => {
                           addNewTimeEntry(entry.description, entry.projectId, entry.taskId).then(() =>
@@ -156,6 +161,7 @@ export default function Main() {
 function NewEntry({ updateTimeEntries }: { updateTimeEntries: () => void }) {
   const { config } = useConfig();
   const [projects, setProjects] = useState<Project[]>([]);
+  const [tags, setTags] = useState<Tag[]>([]);
   const [tasks, setTasks] = useState<Project[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const { pop } = useNavigation();
@@ -163,20 +169,29 @@ function NewEntry({ updateTimeEntries }: { updateTimeEntries: () => void }) {
   useEffect(() => {
     if (isEmpty(config)) return;
 
-    async function getAllProjectsOnWorkspace(): Promise<void> {
+    async function getAllProjectsAndTagsOnWorkspace(): Promise<void> {
       setIsLoading(true);
 
-      const storedProjects: string | undefined = await LocalStorage.getItem("projects");
+      const [storedProjects, storedTags] = await Promise.all([
+        LocalStorage.getItem<string>("projects"),
+        LocalStorage.getItem<string>("tags")
+      ])
       if (storedProjects) setProjects(JSON.parse(storedProjects));
-
-      const { data } = await fetcher(`/workspaces/${config.workspaceId}/projects?page-size=1000&archived=false`);
-
-      setProjects(data || []);
-      LocalStorage.setItem("projects", JSON.stringify(data));
+      if (storedTags) setTags(JSON.parse(storedTags));
+      
+      const [projectsResponse, tagsResponse] = await Promise.all([
+        fetcher(`/workspaces/${config.workspaceId}/projects?page-size=1000&archived=false`),
+        fetcher(`/workspaces/${config.workspaceId}/tags?page-size=1000&archived=false`)
+      ])
+      
+      setProjects(projectsResponse.data || []);
+      LocalStorage.setItem("projects", JSON.stringify(projectsResponse.data));
+      setTags(tagsResponse.data || []);
+      LocalStorage.setItem("tags", JSON.stringify(tagsResponse.data));
       setIsLoading(false);
     }
 
-    getAllProjectsOnWorkspace();
+    getAllProjectsAndTagsOnWorkspace();
   }, [config]);
 
   return (
@@ -187,9 +202,9 @@ function NewEntry({ updateTimeEntries }: { updateTimeEntries: () => void }) {
         <ActionPanel>
           <Action.SubmitForm
             title="Start"
-            onSubmit={({ description, projectId, taskId }) => {
+            onSubmit={({ description, projectId, taskId, tagIds }) => {
               if (description && projectId) {
-                addNewTimeEntry(description, projectId, taskId === "-1" ? null : taskId).then(() =>
+                addNewTimeEntry(description, projectId, taskId === "-1" ? null : taskId, tagIds).then(() =>
                   updateTimeEntries(),
                 );
                 pop();
@@ -237,7 +252,7 @@ function NewEntry({ updateTimeEntries }: { updateTimeEntries: () => void }) {
       {tasks.length ? (
         <Form.Dropdown id="taskId" title="Task">
           <Form.Dropdown.Section>
-            <Form.Dropdown.Item key={-1} value={"-1"} title={"Without task"} icon={{ source: Icon.TextDocument }} />
+            <Form.Dropdown.Item key={-1} value={"-1"} title={"Without task"} icon={{ source: Icon.BlankDocument }} />
           </Form.Dropdown.Section>
 
           <Form.Dropdown.Section title="Project tasks">
@@ -246,14 +261,18 @@ function NewEntry({ updateTimeEntries }: { updateTimeEntries: () => void }) {
                 key={task.id}
                 value={task.id}
                 title={task.name}
-                icon={{ source: Icon.TextDocument }}
+                icon={{ source: Icon.BlankDocument }}
               />
             ))}
           </Form.Dropdown.Section>
         </Form.Dropdown>
       ) : null}
 
-      <Form.TextField id="description" defaultValue="" title="Description" />
+      <Form.TextField id="description" defaultValue="" title="Description" placeholder="What are you working on?" />
+
+      <Form.TagPicker title="Tags" id="tagIds" placeholder="Search tags">
+        {tags.map(tag => <Form.TagPicker.Item key={tag.id} title={tag.name} value={tag.id} />)}
+      </Form.TagPicker>
     </Form>
   );
 }
@@ -306,6 +325,7 @@ async function addNewTimeEntry(
   description: string,
   projectId: string,
   taskId: string | undefined | null,
+  tagIds=[]
 ): Promise<void> {
   showToast(Toast.Style.Animated, "Starting…");
 
@@ -322,6 +342,7 @@ async function addNewTimeEntry(
         end: null,
         duration: null,
       },
+      tagIds,
       customFieldValues: [],
     },
   });